requirements.txt
# Web Framework
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# AI and ML
google-cloud-speech>=2.21.0
google-cloud-texttospeech>=2.16.3
google-generativeai>=0.3.0
openai>=1.3.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
rembg>=2.0.50
numpy>=1.24.0

# Audio Processing
pyaudio>=0.2.11
wave>=0.0.2
io>=0.0.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
typing-extensions>=4.7.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0